extract `QueryBuilder` to a dedicated module.
