+++ datafusion/core/tests/fuzz_cases/aggregation_fuzzer/mod.rs
@@ -43,6 +43,7 @@ use datafusion_common::error::Result;
 mod context_generator;
 mod data_generator;
 mod fuzzer;
+pub mod query_builder;
 
 pub use crate::fuzz_cases::record_batch_generator::ColumnDescr;
 pub use data_generator::DatasetGeneratorConfig;
